split e2e test
